processing of emotes moved to LoadEmotesByPointersSystem
